+++ crates/full-node/sov-db/src/schema/types.rs
@@ -245,3 +245,4 @@ u64_wrapper!(BatchNumber);
 u64_wrapper!(TxNumber);
 u64_wrapper!(EventNumber);
 u64_wrapper!(ProofUniqueId);
+u64_wrapper!(StfInfoUniqueId);